fix unit test

+++ config/config_test.go
@@ -239,7 +239,7 @@ spilled-file-encryption-method = "plaintext"
 	c.Assert(conf.TiKVClient.CommitTimeout, Equals, "41s")
 	c.Assert(conf.TiKVClient.AsyncCommit.Enable, Equals, true)
 	c.Assert(conf.TiKVClient.AsyncCommit.KeysLimit, Equals, uint(123))
-	c.Assert(conf.TiKVClient.AsyncCommit.TotalKeySizeLimit, Equals, uint(1024))
+	c.Assert(conf.TiKVClient.AsyncCommit.TotalKeySizeLimit, Equals, uint64(1024))
 	c.Assert(conf.TiKVClient.MaxBatchSize, Equals, uint(128))
 	c.Assert(conf.TiKVClient.RegionCacheTTL, Equals, uint(6000))
 	c.Assert(conf.TiKVClient.StoreLimit, Equals, int64(0))
